--- client/src/features/EventManage/components/index.ts
@@ -1,3 +1,0 @@
-export { EventInfoSection } from './EventInfoSection';
-export { GuestManageSection } from './GuestManageSection';
-export { GuestModal } from './GuestModal';
