--- client/src/features/EventManage/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useEventManage } from './useEventManage';
