--- client/src/features/EventManage/index.ts
@@ -1,7 +1,0 @@
-export { EventManage } from './pages/EventManage';
-
-export { EventInfoSection, GuestManageSection, GuestModal } from './components';
-
-export { useEventManage } from './hooks';
-
-export type { EventInfo, Guest, EventManageData } from './types';
